requirements.txt
sentence-transformers==2.3.1
langchain
flask
streamlit
langchain_groq
pypdf
python-dotenv
google-api-python-client
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_experimental
langchain-huggingface
langchain-classic
-e .
